Tests for ast-to-llvm.
We don't have a lot of tests, but we have enough tests to know that our
conversion in ast-to-llvm is doing what we expect. Our input language
seems to look like...

(program
  (assign tmp.1 8)
  (result tmp.1))

or

(program
  (assign tmp.1 (+ 3 5))
  (result tmp.1))

which would expand to

%tmp.1 = add i32 3, 5
%result = %tmp.1

which we'll wrap up in a main() in a later pass (wrap-out or something).