Adding in lang0.
So, lets try this.

Although it will be a bit more... er, tracking on our part, I'm going to suggest that we make different languages explicit. This could become problematic. (I'm trying to remember if this was a problem in a previous compiler project.)

Put simply, lang0 is defined by the verifier for lang0 (and, we should write a grammar shortly). When we work backwards, we'll want to then define lang1, and eventually lang2, and so on. The nice thing about this is that our verifiers will be in terms of lang0 structures, so if we mess up (and fail to do a conversion somewhere), then we'll know, because we'll have a lang1 data structure hanging around.

In other words, even though Racket is a dynamically typed language, we can introduce some types into the compiler by having different data structures for different parts of the compiler.

Let me know if this seems like too much bookkeeping.