Some gross things to make stuff work.
We learned things.

1. You can't use constants inline in LLVM's IR. At least, it seems that way.

Constants can be declared as

@t1 = global i32 3

which gives us a constant that can be referenced, as a pointer, anywhere.

Unfortunately, you can't just "use" it in a function call that expects an
int32. At least, we haven't figured out how, yet. (We haven't researched this,
though.)

2. Because of SSA, we'll need to allocate, store, and load values if we don't
use globals for constants.

%t1 = alloca i32

store i32 3, i32* %t1, align 4

%0 = load i32* %t1, align 4

That's one way to handle constants in our language... allocate and store them
at the front, and then load them into temporaries as needed. We could do this.

3. We can also load constants

%t1 = load i32* @t1, align 4

That would be another way to use the constants... load them locally after
declaring them globally.

4. We could use a really gross identity function.

This is what we did in this current version to make things work "quickly." That
is, it "just works," but it is really, really gross. There is no way we could
teach this in a compilers class. Better to go through, extract the constants,
and load them as needed. That, or do it all locally... but, not this. Mercy.

===

We have a driver that will spit out code and we can compile/run it. It seems to
work. But... that's just proof that we have a back-end that generates code that
works. Now, we need to revisit what the "last few passes" look like, because we
probably don't actually have lang0. We probably have "lang0+n", and now need to
get down to the ACTUAL lang0.

If that makes sense.