No more integers.
When we design L1, it will have a list of constants. Because we cannot directly
use an integer constant (or any other, for that matter) in the IR for LLVM, we
will always be operating on names (variables) everywhere. We do need the
"load-const" form, but we no longer need an "int" structure.

This breaks almost all the code we have. No tests work now, nor do any of the
conversion passes. Oops.

But, we'll get there.